[REF] component: use a status number instead of flags
With this commit, we introduce the status internal number, which
represents the current state of the component. It can currently have 4
values:

- CREATED (component is created and not rendered yet)
- RENDERED (component has been rendered at least once, and is not in the
DOM)
- MOUNTED (component has been rendered, and is in the DOM
- DESTROYED

This status number replaces the isMounted and isDestroyed boolean flags.
It has the advantage of making sure that the component is in a
consistent state (it is no longer possible to be destroyed and mounted,
for example)

Another advantage is that it gives us an easy way to track the fact that
a component has been rendered, but is not in the DOM.  This is a subtle
situation where some various events can happen, and we need to be able
to react to that case.